Add support for Mysql and Codendi

+++ src/plugins/mediawiki/bin/create_links.sh
@@ -7,9 +7,24 @@ plugindir=`dirname $absolutedir`
 LINKS=$plugindir/packaging/links/plugin-mediawiki
 cat $LINKS | while read src dest
 do
-	if [ ! -e /$dest ]
+	if [ ! -e /$src ]
 	then
-		echo "Symlinking /$dest -> /$src"
-		ln -s /$src /$dest
+		newsrc=`echo $src | sed 's/mediawiki/mediawiki115/'`
+		if [ -e "/$newsrc" ]
+		then
+			src=$newsrc
+		fi
+	fi
+	if [ -e "/$src" ]
+	then
+		if [ -e /usr/share/codendi ]
+		then
+			dest=`echo $dest | sed 's/gforge/codendi/'`
+		fi
+		if [ ! -e /$dest ]
+		then
+			echo "Symlinking /$dest -> /$src"
+			ln -s /$src /$dest
+		fi
 	fi
 done
